+++ spec/luna_park/extensions/has_errors_spec.rb
@@ -2,11 +2,17 @@
 
 require 'luna_park/extensions/has_errors'
 
+class CustomErrorBase < LunaPark::Errors::Business
+  notify :debug
+  message 'Default message', i18n_key: 'errors.default_key'
+end
+
 class Service
   include LunaPark::Extensions::HasErrors
 
   business_error :business_error
   system_error :system_error
+  custom_error :custom_error, CustomErrorBase
 end
 
 module LunaPark
@@ -16,17 +22,17 @@ module LunaPark
         expect(Service::BusinessError < Errors::Business).to eq true
       end
 
-      context 'when define custom params' do
+      context 'with params' do
         before do
-          Service.business_error(:custom_business_error, 'Custom error', i18n_key: 'errors.custom_error', notify: :info)
+          Service.business_error(:business_error_example, 'Business error', i18n_key: 'errors.business_error', notify: :info)
         end
 
-        subject(:custom_error) { Service::CustomBusinessError }
+        subject(:business_error) { Service::BusinessErrorExample }
 
-        it 'should define all custom params' do
-          expect(custom_error.default_message_block.call({})).to eq 'Custom error'
-          expect(custom_error.i18n_key).to eq 'errors.custom_error'
-          expect(custom_error.default_notify).to eq :info
+        it 'assigns params to the error' do
+          expect(business_error.default_message_block.call({})).to eq 'Business error'
+          expect(business_error.i18n_key).to eq 'errors.business_error'
+          expect(business_error.default_notify).to eq :info
         end
       end
     end
@@ -53,17 +59,17 @@ module LunaPark
         expect(Service::SystemError < Errors::System).to eq true
       end
 
-      context 'when define custom params' do
+      context 'with params' do
         before do
-          Service.business_error(:custom_system_error, 'Custom error', i18n_key: 'errors.custom_error', notify: :info)
+          Service.business_error(:system_error_example, 'System error', i18n_key: 'errors.system_error', notify: :info)
         end
 
-        subject(:custom_error) { Service::CustomSystemError }
+        subject(:system_error) { Service::SystemErrorExample }
 
-        it 'should define all custom params' do
-          expect(custom_error.default_message_block.call({})).to eq 'Custom error'
-          expect(custom_error.i18n_key).to eq 'errors.custom_error'
-          expect(custom_error.default_notify).to eq :info
+        it 'assigns params to the error' do
+          expect(system_error.default_message_block.call({})).to eq 'System error'
+          expect(system_error.i18n_key).to eq 'errors.system_error'
+          expect(system_error.default_notify).to eq :info
         end
       end
     end
@@ -81,6 +87,46 @@ module LunaPark
             an_instance_of(Service::SystemError)
               .and(having_attributes(message: 'Error message', details: { detail: :foo }, notify_lvl: :debug))
           )
+        end
+      end
+    end
+
+    describe '.custom_error' do
+      it 'should define custom error' do
+        expect(Service::CustomError < CustomErrorBase).to eq true
+      end
+
+      context 'with params' do
+        before do
+          Service.custom_error(:custom_error_example, CustomErrorBase, 'Custom error', i18n_key: 'errors.custom_error', notify: :info)
+        end
+
+        subject(:custom_error) { Service::CustomErrorExample }
+
+        it 'assigns params to the error' do
+          expect(custom_error.default_message_block.call({})).to eq 'Custom error'
+          expect(custom_error.i18n_key).to eq 'errors.custom_error'
+          expect(custom_error.default_notify).to eq :info
+        end
+      end
+
+      context 'without params' do
+        before do
+          Service.custom_error(:custom_error_example, CustomErrorBase)
+        end
+
+        subject(:custom_error) { Service::CustomError }
+
+        it 'inherits params from the base error' do
+          expect(custom_error.default_message_block.call({})).to eq 'Default message'
+          expect(custom_error.i18n_key).to eq 'errors.default_key'
+          expect(custom_error.default_notify).to eq :debug
+        end
+      end
+
+      context 'when inherit_from is not a superclass of Errors::Base' do
+        it 'raises error' do
+          expect { Service.custom_error(:custom_error_example, RuntimeError) }.to raise_error(ArgumentError)
         end
       end
     end