+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.5] - 2022-09-27
+Changed
+- Added `.custom_error` method to `Extensions::HasError` to define errors with a custom superclass
+- Added `#inject` method to `Extensions::Injector` - dependencies setter that allows you to create method chains
+
 ## [0.11.4] - 2022-07-11
 Changed
 - allow require Notifiers::Sentry when sentry-ruby version >= 4